Organize into files based on related structs
