+++ node_visitor.go
@@ -1,0 +1,5 @@
+package main
+
+type NodeVisitor interface {
+	Visit(n *Node, depth int)
+}
